requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
langchain-groq>=0.1.0
python-dotenv>=0.19.0
plotly>=5.0.0
pandas>=1.3.0
pydantic>=1.8.0
numpy>=1.21.0
scikit-learn>=1.0.0
torch>=1.9.0
pyverilog>=1.3.0
pytest>=6.0.0
pytest-asyncio>=0.15.0
black>=21.0.0
isort>=5.0.0
mypy>=0.910
